+++ src/engine/documenturldb.cpp
@@ -36,13 +36,13 @@ DocumentUrlDB::~DocumentUrlDB()
     mdb_dbi_close(mdb_txn_env(m_txn), m_dbi);
 }
 
-void DocumentUrlDB::put(uint docId, const QByteArray& url)
+void DocumentUrlDB::put(quint64 docId, const QByteArray& url)
 {
     Q_ASSERT(docId > 0);
     Q_ASSERT(!url.isEmpty());
 
     MDB_val key;
-    key.mv_size = sizeof(uint);
+    key.mv_size = sizeof(quint64);
     key.mv_data = static_cast<void*>(&docId);
 
     MDB_val val;
@@ -56,12 +56,12 @@ void DocumentUrlDB::put(uint docId, const QByteArray& url)
     Q_ASSERT_X(rc == 0, "DocumentUrlDB::put", mdb_strerror(rc));
 }
 
-QByteArray DocumentUrlDB::get(uint docId)
+QByteArray DocumentUrlDB::get(quint64 docId)
 {
     Q_ASSERT(docId > 0);
 
     MDB_val key;
-    key.mv_size = sizeof(uint);
+    key.mv_size = sizeof(quint64);
     key.mv_data = static_cast<void*>(&docId);
 
     MDB_val val;
@@ -74,12 +74,12 @@ QByteArray DocumentUrlDB::get(uint docId)
     return QByteArray::fromRawData(static_cast<char*>(val.mv_data), val.mv_size);
 }
 
-void DocumentUrlDB::del(uint docId)
+void DocumentUrlDB::del(quint64 docId)
 {
     Q_ASSERT(docId > 0);
 
     MDB_val key;
-    key.mv_size = sizeof(uint);
+    key.mv_size = sizeof(quint64);
     key.mv_data = static_cast<void*>(&docId);
 
     int rc = mdb_del(m_txn, m_dbi, &key, 0);
